Move makePromise to a separate file
